Add a destiny 2 page
