+++ src/app/activities/activities.module.js
@@ -1,11 +1,10 @@
 import angular from 'angular';
-import 'angular-duration-format';
 
 import bungieApiModule from '../bungie-api/bungie-api.module';
 import { ActivitiesComponent } from './activities.component';
 
 export default angular
-  .module('activitiesModule', ['angular-duration-format', bungieApiModule])
+  .module('activitiesModule', [bungieApiModule])
   .component('activities', ActivitiesComponent)
   .config(($stateProvider) => {
     'ngInject';
