+++ src/app/loadout-builder/loadout-builder-item.component.js
@@ -84,6 +84,5 @@ function LoadoutBuilderItemCtrl($scope, $element, ngDialog, dimStoreService) {
       }
     }
   });
-
 }
